+++ test/bonny/watcher/impl_test.exs
@@ -3,14 +3,6 @@ defmodule Bonny.Watcher.ImplTest do
   use ExUnit.Case, async: true
   alias Bonny.Watcher.Impl
 
-  setup do
-    bypass = Bypass.open()
-    k8s_config = K8s.Conf.from_file("test/support/kubeconfig.yaml")
-    k8s_config = %{k8s_config | url: "http://localhost:#{bypass.port}/"}
-
-    {:ok, bypass: bypass, k8s_config: k8s_config}
-  end
-
   defmodule Whizbang do
     @moduledoc false
     def add(evt), do: emit(:added, evt)
@@ -23,77 +15,70 @@ defmodule Bonny.Watcher.ImplTest do
     end
   end
 
-  def added_chunk do
-    "{\"type\":\"ADDED\",\"object\":{\"apiVersion\":\"example.com/v1\",\"kind\":\"Widget\",\"metadata\":{\"annotations\":{\"kubectl.kubernetes.io/last-applied-configuration\":\"{\\\"apiVersion\\\":\\\"example.com/v1\\\",\\\"kind\\\":\\\"Widget\\\",\\\"metadata\\\":{\\\"annotations\\\":{},\\\"name\\\":\\\"test-widget\\\",\\\"namespace\\\":\\\"default\\\"}}\\n\"},\"clusterName\":\"\",\"creationTimestamp\":\"2018-12-17T06:26:41Z\",\"generation\":1,\"name\":\"test-widget\",\"namespace\":\"default\",\"resourceVersion\":\"705460\",\"selfLink\":\"/apis/example.com/v1/namespaces/default/widgets/test-widget\",\"uid\":\"b7464e30-01c4-11e9-9066-025000000001\"}}}\n"
-  end
-
-  def deleted_chunk do
-    "{\"type\":\"DELETED\",\"object\":{\"apiVersion\":\"example.com/v1\",\"kind\":\"Widget\",\"metadata\":{\"annotations\":{\"kubectl.kubernetes.io/last-applied-configuration\":\"{\\\"apiVersion\\\":\\\"example.com/v1\\\",\\\"kind\\\":\\\"Widget\\\",\\\"metadata\\\":{\\\"annotations\\\":{},\\\"name\\\":\\\"test-widget\\\",\\\"namespace\\\":\\\"default\\\"}}\\n\"},\"clusterName\":\"\",\"creationTimestamp\":\"2018-12-17T06:26:41Z\",\"generation\":1,\"name\":\"test-widget\",\"namespace\":\"default\",\"resourceVersion\":\"705464\",\"selfLink\":\"/apis/example.com/v1/namespaces/default/widgets/test-widget\",\"uid\":\"b7464e30-01c4-11e9-9066-025000000001\"}}}\n"
-  end
-
   describe "new/1" do
     test "returns the default state" do
-      assert %Impl{mod: Widget, spec: %Bonny.CRD{}, config: %K8s.Conf{}} = Impl.new(Widget)
+      assert %Impl{controller: Widget, spec: %Bonny.CRD{}} = Impl.new(Widget)
     end
   end
 
-  describe "get_resource_version/2" do
-    test "returns the kubernetes resource version", %{bypass: bypass, k8s_config: k8s_config} do
-      Bypass.expect_once(bypass, fn conn ->
-        assert conn.method == "GET"
-        assert conn.request_path == "/apis/example.com/v1/namespaces/default/widgets"
-        Plug.Conn.resp(conn, 200, ~s<{"metadata": {"resourceVersion": "1337"}}>)
-      end)
+  describe "set_resource_version/2" do
+    test "sets the resource version from a watch event" do
+      event = %{
+        "object" => %{"metadata" => %{"resourceVersion" => "3"}}
+      }
 
       state = Impl.new(Widget)
-      state = %{state | config: k8s_config}
-      {:ok, %{resource_version: rv}} = Impl.get_resource_version(state)
+      new_state = Impl.set_resource_version(state, event)
+      assert new_state.resource_version == 3
+    end
 
-      assert rv == "1337"
+    test "sets the resource version" do
+      state = Impl.new(Widget)
+      state = Map.put(state, :resource_version, 1)
+      new_state = Impl.set_resource_version(state, 3)
+      assert new_state.resource_version == 3
+    end
+
+    test "does not decrease the resource version" do
+      state = Impl.new(Widget)
+      state = Map.put(state, :resource_version, 200)
+      new_state = Impl.set_resource_version(state, 3)
+      assert new_state.resource_version == 200
     end
   end
 
   describe "watch_for_changes/2" do
-    test "returns changes to a CRD resource", %{bypass: bypass, k8s_config: k8s_config} do
-      added = added_chunk()
-      deleted = deleted_chunk()
-
-      Bypass.expect_once(bypass, fn conn ->
-        assert conn.method == "GET"
-        assert conn.request_path == "/apis/example.com/v1/namespaces/default/widgets"
-        assert conn.query_string == "resourceVersion=1337&watch=true"
-
-        conn = Plug.Conn.send_chunked(conn, 200)
-        {:ok, conn} = Plug.Conn.chunk(conn, added)
-        {:ok, conn} = Plug.Conn.chunk(conn, deleted)
-        conn
-      end)
+    test "returns changes to a CRD resource" do
+      added = Bonny.K8sMockClient.added_chunk()
+      deleted = Bonny.K8sMockClient.deleted_chunk()
 
       state = Impl.new(Widget)
-      state = %{state | config: k8s_config}
-      state = %{state | resource_version: "1337"}
+      state = %{state | resource_version: 1337}
 
       Impl.watch_for_changes(state, self())
 
-      assert_receive %HTTPoison.AsyncStatus{code: 200}, 1_000
-      assert_receive %HTTPoison.AsyncHeaders{}, 1_000
       assert_receive %HTTPoison.AsyncChunk{chunk: ^added}, 1_000
       assert_receive %HTTPoison.AsyncChunk{chunk: ^deleted}, 1_000
-      assert_receive %HTTPoison.AsyncEnd{}, 1_000
     end
   end
 
   describe "parse_chunk/1" do
     test "strips whitespace and parses json" do
-      result = added_chunk() |> Impl.parse_chunk()
+      chunk =
+        "{\"type\":\"ADDED\",\"object\":{\"apiVersion\":\"example.com/v1\",\"kind\":\"Widget\",\"metadata\":{\"annotations\":{\"kubectl.kubernetes.io/last-applied-configuration\":\"{\\\"apiVersion\\\":\\\"example.com/v1\\\",\\\"kind\\\":\\\"Widget\\\",\\\"metadata\\\":{\\\"annotations\\\":{},\\\"name\\\":\\\"test-widget\\\",\\\"namespace\\\":\\\"default\\\"}}\\n\"},\"clusterName\":\"\",\"creationTimestamp\":\"2018-12-17T06:26:41Z\",\"generation\":1,\"name\":\"test-widget\",\"namespace\":\"default\",\"resourceVersion\":\"705460\",\"selfLink\":\"/apis/example.com/v1/namespaces/default/widgets/test-widget\",\"uid\":\"b7464e30-01c4-11e9-9066-025000000001\"}}}\n"
+
+      result = Impl.parse_chunk(chunk)
       assert %{"type" => "ADDED"} = result
     end
   end
 
   describe "dispatch/2" do
     test "dispatches a kubernetes API event to the given module" do
-      added_event = added_chunk() |> Impl.parse_chunk()
-      Impl.dispatch(added_event, Whizbang)
+      chunk =
+        "{\"type\":\"ADDED\",\"object\":{\"apiVersion\":\"example.com/v1\",\"kind\":\"Widget\",\"metadata\":{\"annotations\":{\"kubectl.kubernetes.io/last-applied-configuration\":\"{\\\"apiVersion\\\":\\\"example.com/v1\\\",\\\"kind\\\":\\\"Widget\\\",\\\"metadata\\\":{\\\"annotations\\\":{},\\\"name\\\":\\\"test-widget\\\",\\\"namespace\\\":\\\"default\\\"}}\\n\"},\"clusterName\":\"\",\"creationTimestamp\":\"2018-12-17T06:26:41Z\",\"generation\":1,\"name\":\"test-widget\",\"namespace\":\"default\",\"resourceVersion\":\"705460\",\"selfLink\":\"/apis/example.com/v1/namespaces/default/widgets/test-widget\",\"uid\":\"b7464e30-01c4-11e9-9066-025000000001\"}}}\n"
+
+      added_chunk = Impl.parse_chunk(chunk)
+      Impl.dispatch(added_chunk, Whizbang)
       assert_received({:added, added_event})
     end
   end