+++ test/test_helper.exs
@@ -1,2 +1,1 @@
 ExUnit.start()
-Application.ensure_all_started(:bypass)